+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 This file contains the list of changes made to pytation.
 
 
+## 0.0.3
+
+2021 June 28
+
+*   Refactored and improved station validate.
+*   Added test "--include" and "--exclude" command-line options.
+
+
+## 0.0.2
+
+2021 June 25
+
+*   Fixed installation dependencies.
+
+
 ## 0.0.1
 
 2021 June 25